# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "description": "`{ \"error\": … }`: rejections, refused ceremonies, rendered store errors.",
        "properties": {
          "error": {
            "description": "What went wrong, in words meant for the end user.",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "EmailNextForm": {
        "properties": {
          "email": {
            "description": "The address to send the link to.",
            "type": "string"
          },
          "next": {
            "description": "Where to send the browser afterwards; must be a local path.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
      "FlowError": {
        "description": "A flow step was refused. `next` is where a browser would have been sent.",
        "properties": {
          "error": {
            "description": "Why the step was refused.",
            "type": "string"
          },
          "next": {
            "description": "Where a browser would have been redirected to.",
            "type": "string"
          }
        },
        "required": [
          "error",
          "next"
        ],
        "type": "object"
      },
      "FlowResult": {
        "description": "A flow step succeeded. Browsers get a redirect to `next`; JSON clients\nget this.",
        "properties": {
          "message": {
            "description": "A note worth showing the user, when the step produced one.",
            "type": [
              "string",
              "null"
            ]
          },
          "next": {
            "description": "Where a browser would have been redirected to.",
            "type": "string"
          }
        },
        "required": [
          "next"
        ],
        "type": "object"
      },
...
      "MfaCodeForm": {
        "description": "One form serves both steps: without `code` it requests a code to be sent,\nwith `code` it verifies it. An empty `code` counts as absent.",
        "properties": {
          "code": {
            "description": "The code from the message; omit it to have one sent.",
            "type": [
              "string",
              "null"
            ]
          },
          "next": {
            "description": "Where to send the browser afterwards; must be a local path.",
            "type": [
              "string",
              "null"
            ]
          },
          "trust_device": {
            "description": "Any non-empty value marks this browser as trusted.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "session_cookie": {
        "in": "cookie",
        "name": "authery-session-id",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Authery",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/login/email": {
      "get": {
        "operationId": "login_email_callback",
        "parameters": [
          {
            "description": "The single-use challenge code carried by the emailed link.",
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "description": "The single-use challenge code carried by the emailed link.",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResult"
                }
              }
            },
            "description": "JSON clients: the step succeeded.",
            "headers": {
              "X-Auth-Token": {
                "description": "Session token for `Authorization: Bearer`, present when this request established a new session.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "Browser clients: redirect to the next page, with `?error=` on refusal.",
            "headers": {
              "X-Auth-Token": {
                "description": "Session token for `Authorization: Bearer`, present when this request established a new session.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowError"
                }
              }
            },
            "description": "JSON clients: the step was refused."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The store refused the request and chose to say why."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Finish a login link",
        "tags": [
          "Email"
        ]
      },
      "post": {
        "operationId": "login_email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailNextForm"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/EmailNextForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResult"
                }
              }
            },
            "description": "JSON clients: the step succeeded."
          },
          "303": {
            "description": "Browser clients: redirect to the next page, with `?error=` on refusal."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowError"
                }
              }
            },
            "description": "JSON clients: the step was refused."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The store refused the request and chose to say why."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Send a login link",
        "tags": [
          "Email"
        ]
      }
    },
    "/login/mfa/otp": {
      "post": {
        "operationId": "login_mfa_otp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeForm"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResult"
                }
              }
            },
            "description": "JSON clients: the step succeeded.",
            "headers": {
              "X-Auth-Token": {
                "description": "Session token for `Authorization: Bearer`, present when this request established a new session.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "Browser clients: redirect to the next page, with `?error=` on refusal.",
            "headers": {
              "X-Auth-Token": {
                "description": "Session token for `Authorization: Bearer`, present when this request established a new session.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowError"
                }
              }
            },
            "description": "JSON clients: the step was refused."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The store refused the request and chose to say why."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Complete MFA with an emailed code",
        "tags": [
          "MFA"
        ]
      }
    },
    "/login/mfa/recovery": {
      "post": {
        "operationId": "login_mfa_recovery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyForm"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResult"
                }
              }
            },
            "description": "JSON clients: the step succeeded.",
            "headers": {
              "X-Auth-Token": {
                "description": "Session token for `Authorization: Bearer`, present when this request established a new session.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "Browser clients: redirect to the next page, with `?error=` on refusal.",
            "headers": {
              "X-Auth-Token": {
                "description": "Session token for `Authorization: Bearer`, present when this request established a new session.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowError"
                }
              }
            },
            "description": "JSON clients: the step was refused."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The store refused the request and chose to say why."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Complete MFA with a recovery code",
        "tags": [
          "MFA"
        ]
      }
    },
    "/login/mfa/sms": {
      "post": {
        "operationId": "login_mfa_sms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeForm"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResult"
                }
              }
            },
            "description": "JSON clients: the step succeeded.",
            "headers": {
              "X-Auth-Token": {
                "description": "Session token for `Authorization: Bearer`, present when this request established a new session.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "Browser clients: redirect to the next page, with `?error=` on refusal.",
            "headers": {
              "X-Auth-Token": {
                "description": "Session token for `Authorization: Bearer`, present when this request established a new session.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowError"
                }
              }
            },
            "description": "JSON clients: the step was refused."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The store refused the request and chose to say why."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Complete MFA with a texted code",
        "tags": [
          "MFA"
        ]
      }
    },
    "/login/mfa/totp": {
      "post": {
        "operationId": "login_mfa_totp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyForm"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResult"
                }
              }
            },
            "description": "JSON clients: the step succeeded.",
            "headers": {
              "X-Auth-Token": {
                "description": "Session token for `Authorization: Bearer`, present when this request established a new session.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "303": {
            "description": "Browser clients: redirect to the next page, with `?error=` on refusal.",
            "headers": {
              "X-Auth-Token": {
                "description": "Session token for `Authorization: Bearer`, present when this request established a new session.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowError"
                }
              }
            },
            "description": "JSON clients: the step was refused."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The store refused the request and chose to say why."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Complete MFA with an authenticator code",
        "tags": [
          "MFA"
        ]
      }
    },
...
  }
}